Create dns-zones-creator.py
*LICENSE EXCEPTION : this file despite continued further modification by me in my learning process is mostly from professor Mike O'Leary (Townsend / "Cyber Operations...") and should be treated as such.  This is here to demonstrate my learning and reflection of common dns forward and reverse setup scripting, and to experiment with getting out of 'the bash comfort zone'

The ultimate goal would be to be able to reproduce or more importantly, similarly productive scripts, spontaneously, with little guide

This was a fun lab exercize, and great for learning the dns/BIND service in depth

Mike O'Leary states in "Cyber Operations,..." "This program reads each line from stars.csv, storing the host name and the IP address. It builds the corresponding FQDN and uses that to write the matching address record line to forward.txt. It splits the IP address up into octets, reverses them adding “.in-addr.arpa.” to build the pointer record, which is written to reverse.txt. The resulting files can then be copied and pasted into appropriate zone files.   ...The advantage of this approach is that it avoids typographical errors and ensures consistency between the data in the forward and reverse zone files."